+++ packages/semaphore-identity-pcd/src/SemaphoreIdentityPCD.ts
@@ -1,6 +1,9 @@
-import { PCD, PCDPackage } from "@pcd/pcd-types";
+import { PCD, PCDPackage, SerializedPCD } from "@pcd/pcd-types";
 import { Identity } from "@semaphore-protocol/identity";
 import JSONBig from "json-bigint";
+import { v4 as uuid } from "uuid";
+
+export const SemaphoreIdentityPCDTypeName = "semaphore-identity-pcd";
 
 export interface SemaphoreIdentityPCDArgs {
   identity: Identity;
@@ -15,38 +18,46 @@ export type SemaphoreIdentityPCDProof = undefined;
 export class SemaphoreIdentityPCD
   implements PCD<SemaphoreIdentityPCDClaim, SemaphoreIdentityPCDProof>
 {
-  type = "SemaphoreIdentityPCD";
+  type = SemaphoreIdentityPCDTypeName;
   claim: SemaphoreIdentityPCDClaim;
   proof: SemaphoreIdentityPCDProof;
+  id: string;
 
-  public constructor(claim: SemaphoreIdentityPCDClaim) {
+  public constructor(id: string, claim: SemaphoreIdentityPCDClaim) {
     this.claim = claim;
     this.proof = undefined;
+    this.id = id;
   }
 }
 
 export async function prove(
   args: SemaphoreIdentityPCDArgs
 ): Promise<SemaphoreIdentityPCD> {
-  return new SemaphoreIdentityPCD({ identity: args.identity });
+  return new SemaphoreIdentityPCD(uuid(), { identity: args.identity });
 }
 
 export async function verify(pcd: SemaphoreIdentityPCD): Promise<boolean> {
   return pcd?.claim?.identity !== undefined;
 }
 
-export async function serialize(pcd: SemaphoreIdentityPCD): Promise<string> {
-  return JSONBig.stringify({
-    type: pcd.type,
-    identity: pcd.claim.identity.toString(),
-  });
+export async function serialize(
+  pcd: SemaphoreIdentityPCD
+): Promise<SerializedPCD<SemaphoreIdentityPCD>> {
+  return {
+    type: SemaphoreIdentityPCDTypeName,
+    pcd: JSONBig.stringify({
+      type: pcd.type,
+      id: pcd.id,
+      identity: pcd.claim.identity.toString(),
+    }),
+  } as SerializedPCD<SemaphoreIdentityPCD>;
 }
 
 export async function deserialize(
   serialized: string
 ): Promise<SemaphoreIdentityPCD> {
   const parsed = JSONBig.parse(serialized);
-  return new SemaphoreIdentityPCD({
+  return new SemaphoreIdentityPCD(parsed.id, {
     identity: new Identity(parsed.identity),
   });
 }
@@ -60,7 +71,7 @@ export const SemaphoreIdentityPCDPackage: PCDPackage<
   SemaphoreIdentityPCDProof,
   SemaphoreIdentityPCDArgs
 > = {
-  name: "SemaphoreIdentityPCD",
+  name: SemaphoreIdentityPCDTypeName,
   prove,
   verify,
   serialize,