+++ packages/pcd-collection/CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Patch Changes
 
 - Preparing for zuzalu.
+- Updated dependencies
+  - @pcd/pcd-types@0.0.3
 
 ## 0.0.2